+++ CHANGELOG.md
@@ -7,6 +7,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.3] - 2024-04-18
+
+### Added
+- Added filtering on the data to remove duplicated data in patient file. 
+- Added a dashboard to communicate with the Posit database. 
+- Added functionality to show stored GPT4 letter in eval dashboard
+- Added 'refresh' functionality when switching to a new patient in eval dashboard 
+
+
+## [0.4.2] - 2024-04-15
+
+### Added
+- Added endpoint to the API for deleting old generated discharge documents
+- Added unit tests
+
+## [0.4.1] - 2024-04-12
+
+### Changed
+- First working version of the API including database tables
+
+### Added
+- Added unit tests for the API
+
+## [0.4.0] - 2024-04-11
+
+### Changed
+- changed database to the Posit Connect database
+- Updated the department prompts
+- automatic evaluation using GPT first version
+
 ## [0.3.3] - 2024-03-11
 
 ### Added